add cname migration added

+++ node/mcollective/mcollective/agent/migrate-2.1.4.rb
@@ -44,9 +44,11 @@ EOF
       end
       
       mysql_ctl_sh = "#{app_home}/mysql-5.1/#{app_name}_mysql_ctl.sh"
-      output += "Migrating mysql_ctl_sh: #{mysql_ctl_sh}\n"
-      FileUtils.rm_rf mysql_ctl_sh
-      FileUtils.ln_s "#{cartridge_root_dir}/mysql-5.1/info/bin/mysql_ctl.sh", mysql_ctl_sh
+      if File.exists?(mysql_ctl_sh)
+        output += "Migrating mysql_ctl_sh: #{mysql_ctl_sh}\n"
+        FileUtils.rm_rf mysql_ctl_sh
+        FileUtils.ln_s "#{cartridge_root_dir}/embedded/mysql-5.1/info/bin/mysql_ctl.sh", mysql_ctl_sh
+      end
       
     else
       exitcode = 127
